requirements.txt
pytest==7.1.2
python-dotenv==1.0.1
requests==2.32.3
requests-mock==1.12.1
pytest-mock==3.14.0